.bash_completion.d/adb.bash: fix disable-verity bin/my-adb: fix disable-verity

+++ .bash_completion.d/adb.bash
@@ -59,7 +59,15 @@ _adb() {
     fi
 
     OPTIONS="-d -e -s -p"
-    COMMAND="devices connect disconnect push pull sync shell emu logcat lolcat forward jdwp install uninstall bugreport help version start-server kill-server get-state get-serialno status-window remount reboot reboot-bootloader root usb tcpip backup get-devpath ppp restore wait-for-device"
+
+    if test -e ~/.config/system-config/adb-commands; then
+        COMMAND=$(
+            . ~/.config/system-config/adb-commands
+            echo ${!adb_commands[*]}
+               )
+    else
+        COMMAND="devices connect disconnect push pull sync shell emu logcat lolcat forward jdwp install uninstall bugreport help version start-server kill-server get-state get-serialno status-window remount reboot reboot-bootloader root usb tcpip backup get-devpath ppp restore wait-for-device"
+    fi
 
     case $where in
         OPTIONS|OPT_SERIAL|OPT_PATH)
